Code Quality Update
* Moved commonly reused statics to public static variables in class Utils.
* Added exception handling for common, foreseeable user errors
* Added null tests for same purpose
* Added summary documentation for classes and methods
* Fixed nullable declarations in JSON Object classes
* Added aMessages Object for better Commdata deserialization
* Moved all json string trimming to LoadJson method
* Converted void SaveFiles to type bool for errorhandling

+++ Ostranauts Ship Importer/Utils.cs
@@ -4,16 +4,42 @@ using System.Text.Json;
 
 namespace Ostranauts_Ship_Importer
 {
+    /// <summary>
+    /// Contains utility methods for Ostranauts Ship Importer
+    /// </summary>
     internal class Utils
     {
+        public static string saveInfoFile = "saveInfo.json";
+        public static string jsonExtension = ".json";
         private const int FirstIndex = 0;
 
-        // Reads and serializes JSON into an object of type T (ship or saveInfo)
-        public static T LoadJson<T>(string json)
+        /// <summary>
+        /// Reads <paramref name="json"/> and deserializes into a(n) <typeparamref name="T"/> object
+        /// </summary>
+        /// <typeparam name="T">JSON object (ship or saveInfo)</typeparam>
+        /// <param name="json">JSON text to serialize</param>
+        /// <returns> Deserialized <typeparamref name="T"/> object</returns>
+        public static T? LoadJson<T>(string json)
         {
-            return JsonSerializer.Deserialize<T[]>(json)[0];
+            json = json.Trim();
+            try
+            {
+                return JsonSerializer.Deserialize<T[]>(json)[0];
+            }
+            catch
+            {
+                MessageBox.Show("A specified file does not does not contain readable JSON data.", "Error Parsing JSON Data!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                return default(T?);
+            }
         }
-        // Swap the required JSON fields between the two objects, nulls unneeded arrays
+
+        /// <summary>
+        /// Swap the required JSON fields between the two objects, nulls unneeded arrays
+        /// </summary>
+        /// <param name="importShip">JSON object Ship to import</param>
+        /// <param name="replaceShip">JSON object Ship to replace</param>
+        /// <param name="unharmed">No wear and tear flag</param>
+        /// <returns>JSON object Ship with required fields replaced</returns>
         public static Ship SwapShipData(Ship importShip, Ship replaceShip, bool unharmed)
         {
             replaceShip.strName = importShip.strName;
@@ -35,7 +61,13 @@ namespace Ostranauts_Ship_Importer
 
             return replaceShip;
         }
-        // Finds the selected ship's JSON file in the save's zipfile, then calls LoadJson with that data.
+
+        /// <summary>
+        /// Finds the selected ship's JSON file in the save's zipfile, then calls LoadJson with that data.
+        /// </summary>
+        /// <param name="shipFileName">The JSON ship filename to read</param>
+        /// <param name="saveFile">The filename of the save containing <paramref name="shipFileName"/></param>
+        /// <returns></returns>
         public static Ship ReadShipFromSave(string shipFileName, string saveFile)
         {
             using (ZipArchive zip = ZipFile.OpenRead(saveFile))
@@ -45,14 +77,21 @@ namespace Ostranauts_Ship_Importer
                     if (entry.Name == shipFileName)
                     {
                         using (StreamReader sr = new(entry.Open()))
-                            return (Utils.LoadJson<Ship>(sr.ReadToEnd().Trim()));
+                            return (Utils.LoadJson<Ship>(sr.ReadToEnd()));
                     }
                 }
             }
             return null;
         }
-        // Copies the existing save to new folder and updates the selected ship file and the saveInfo JSONs 
-        public static void SaveFiles(string saveFile, string shipName, Ship outShip)
+
+        /// <summary>
+        /// Copies the existing <paramref name="saveFile"/> to a new folder and updates it with modified <paramref name="outShip"/> and the saveInfo JSON files 
+        /// </summary>
+        /// <param name="saveFile">Full name of save file</param>
+        /// <param name="shipName">Full name of ship file</param>
+        /// <param name="outShip">Final modified JSON Ship object</param>
+        /// <returns>True if completed without errors.</returns>
+        public static bool SaveFiles(string saveFile, string shipName, Ship outShip)
         {
             // set up some file/folder names for future use
             string oldFolder = saveFile.Substring(FirstIndex, 1 + saveFile.LastIndexOf('\\'));
@@ -63,20 +102,51 @@ namespace Ostranauts_Ship_Importer
             // set up options for reserialization of JSON from strings
             var options = new JsonSerializerOptions { WriteIndented = true, DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull };
 
-            // Copy save data to new folder 
+            // If the new folder doesn't already exist, copy save data to new folder 
+            if (Directory.Exists(newFolder))
+            {
+                MessageBox.Show("Error creating new save Folder:\n\n" + newFolder + "\n\nThe folder may already exist, or write permission was denied.", "Error Creating Folder!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                return false;
+            }
             CopyDir.Copy(oldFolder, newFolder);
+
             File.Move(newFolder + oldFileName, newFolder + newFileName);
 
             UpdateSaveInfo(newFolder, options);
 
             UpdateZippedFiles(newFolder, newFileName, shipName, outShip, options);
+
+            return true;
         }
 
-        //Open, modify and save saveInfo changes (save name, save log, and time)
+        /// <summary>
+        /// Open, modify and save saveInfo changes (save name, save log, and time) in <paramref name="newFolder"/>.
+        /// </summary>
+        /// <param name="newFolder">Name of new save folder</param>
+        /// <param name="options">Options info for JSON serialation</param>
         public static void UpdateSaveInfo(string newFolder, JsonSerializerOptions options)
         {
+            string? saveInfoData = null;
+            string fullInfoFilename = newFolder + saveInfoFile;
+            SaveInfo? saveInfo = null;
 
-            SaveInfo saveInfo = LoadJson<SaveInfo>(File.ReadAllText(newFolder + "saveInfo.json").Trim());
+            try
+            {
+                saveInfoData = File.ReadAllText(fullInfoFilename);
+            }
+            catch
+            {
+                MessageBox.Show("Error opening file:\n\n" + fullInfoFilename + "\n\nThe file does not exist or could not be accessed.", "Error Reading Save File!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                return;
+            }
+
+            if (saveInfoData != null)
+            {
+                saveInfo = LoadJson<SaveInfo>(saveInfoData);
+            }
+
+            if (saveInfo == null)
+                return;
 
             saveInfo.strName += "_imported_ship";
             saveInfo.strSaveLog += "* " + DateTime.Now.ToString("M/d/yyyy h:mm:ss tt") + " Ship Imported!\n";
@@ -84,10 +154,17 @@ namespace Ostranauts_Ship_Importer
 
             SaveInfo[] saveInfoArray = { saveInfo };
             string output = JsonSerializer.Serialize(saveInfoArray, options);
-            File.WriteAllText(newFolder + "saveInfo.json", output);
+            File.WriteAllText(newFolder + saveInfoFile, output);
         }
 
-        // Opens the new save file's zip to replace the selected ship with the new data
+        /// <summary>
+        /// Opens the new save file's zip to replace the selected ship with the new data
+        /// </summary>
+        /// <param name="newFolder">Name of new save folder</param>
+        /// <param name="newFileName">Name of the new save file</param>
+        /// <param name="shipName">Filename of the modified ship</param>
+        /// <param name="outShip">The modified Ship data</param>
+        /// <param name="options">Options for JSON Serialization</param>
         public static void UpdateZippedFiles(string newFolder, string newFileName, string shipName, Ship outShip, JsonSerializerOptions options)
         {
             using (ZipArchive zip = ZipFile.Open(newFolder + newFileName, ZipArchiveMode.Update))
@@ -95,7 +172,7 @@ namespace Ostranauts_Ship_Importer
 
                 foreach (ZipArchiveEntry entry in zip.Entries)
                 {
-                    if (entry.Name == "saveInfo.json" || entry.Name == shipName)
+                    if (entry.Name == saveInfoFile || entry.Name == shipName)
                     {
                         entry.Delete();
                         break;
@@ -119,25 +196,29 @@ namespace Ostranauts_Ship_Importer
                 }
 
                 // Also replace saveInfo.json in the zip file (why!?)
-                zip.CreateEntryFromFile(newFolder + "saveInfo.json", "saveInfo.json");
-                zip.Dispose();
+                zip.CreateEntryFromFile(newFolder + saveInfoFile, saveInfoFile);
             }
         }
-        // Creates a List of Shipnames to populate combobox
+
+        /// <summary>
+        /// Creates a List of Shipnames read from <paramref name="saveFile"/> to populate combobox
+        /// </summary>
+        /// <param name="saveFile">Full name of existing save file.</param>
+        /// <returns></returns>
         public static List<string> GenerateShipList(string saveFile)
         {
-            List<string> shipList = [];
+            List<string>? shipList = new();
             using (ZipArchive zip = ZipFile.OpenRead(saveFile))
             {
                 foreach (ZipArchiveEntry entry in zip.Entries)
                 {
                     string ship = entry.FullName;
-                    if (ship.StartsWith("ships/") && ship.EndsWith(".json"))
+                    if (ship.StartsWith("ships/") && ship.EndsWith(jsonExtension))
                     {
                         string shortShip = ship.Substring(6);
                         if (shortShip[1] == '-')
                         {
-                            shortShip = shortShip.Replace(".json", "");
+                            shortShip = shortShip.Replace(jsonExtension, "");
                             shipList.Add(shortShip);
                         }
                     }
